requirements.txt
python-telegram-bot>=21.1.1
fastapi==0.109.2
uvicorn[standard]==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
supabase==2.3.4
python-dotenv==1.0.1
aiohttp==3.9.3
apscheduler==3.10.4
pytz==2024.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
itsdangerous==2.1.2
